bugfix: fix race condition in TSO storebuffer thread
RxChannel.recv consumes the item from the queue. This leads to a race
condition in TSO storebuffer when it is first consumes the pending store
from the buffer but later could be scheduled away on memory lock. In
this case, later reads will not find this store nor in store buffer
itself, nor in the memory.

Fix the bug by introducing the new peek method for RxChannel. This
method will just wait for the item, but will not consume it.